[Auto Generated] 2.0.0

+++ Sources/code/application/Models/User/SendOtpSuccessUserAppModel.swift
@@ -8,7 +8,7 @@ public extension ApplicationClient.User {
     */
     class SendOtpSuccess: Codable {
         
-        public var resendTimer: Int?
+        public var resendTimer: Int64?
         
         public var resendToken: String?
         
@@ -45,7 +45,7 @@ public extension ApplicationClient.User {
             
         }
 
-        public init(countryCode: String? = nil, message: String? = nil, mobile: String? = nil, registerToken: String? = nil, requestId: String? = nil, resendTimer: Int? = nil, resendToken: String? = nil, success: Bool? = nil) {
+        public init(countryCode: String? = nil, message: String? = nil, mobile: String? = nil, registerToken: String? = nil, requestId: String? = nil, resendTimer: Int64? = nil, resendToken: String? = nil, success: Bool? = nil) {
             
             self.resendTimer = resendTimer
             
@@ -70,7 +70,7 @@ public extension ApplicationClient.User {
             
             
             do {
-                resendTimer = try container.decode(Int.self, forKey: .resendTimer)
+                resendTimer = try container.decode(Int64.self, forKey: .resendTimer)
             
             } catch DecodingError.typeMismatch(let type, let context) {
                 print("Type '\(type)' mismatch:", context.debugDescription)
